style, minor, and refactor from to target
