+++ composer.json
@@ -20,7 +20,7 @@
       "php" : ">=5.4.0",
       "michaels/data-manager": "dev-develop",
       "ostico/phporient": "dev-master",
-    "brightzone/gremlin-php": "*"
+      "brightzone/gremlin-php": "*"
 
     },
     "require-dev": {
